requirements.txt
tensorflow>=2.15.0
streamlit>=1.28.0
scikit-learn>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0
pillow>=8.0.0
kagglehub>=0.3.0
jupyter>=1.0.0
pandas>=1.5.0
gradio==5.46.0